Fixed Edit/Reply Forms, Only let creator of book modify/delete

+++ public/index.php
@@ -106,12 +106,11 @@ if ($result && $statement->rowCount() > 0) {?>
 						<?php if (escape($row["is_available"] == "Available" && $_SESSION["usertype"] == "Buyer")) {
     echo "<a href=\"buy.php?bookid=" . urlencode($row['book_id']) . "\" class=\"btn btn-info\">Buy Textbook</a>";
 }?>
-<?php if (escape($row["is_available"] == "Available" && (($_SESSION["usertype"] == "Seller") || ($_SESSION["usertype"] == "Admin")))) {
+<?php if (escape(($row["is_available"] == "Available" && $_SESSION["usertype"] == "Seller" && $_SESSION["id"] == $row["book_creator"]) || $_SESSION["usertype"] == "Admin")) {
     echo "<a href=\"edit.php?bookid=" . urlencode($row['book_id']) . "\" class=\"btn btn-info\">Edit Textbook</a>";
     echo "<a href=\"delete.php?bookid=" . urlencode($row['book_id']) . "&bookname=" . urlencode($row['book_name']) . "\" class=\"btn btn-info\">Delete Textbook</a>";
 }?>
-
-
+							</div>
 						</td>
 					</tr>
 				<?php }?>
